Add support for scoped validators
When sharing validators, often the interfaces provided by the objects being
validated are not uniform.

Furthermore, the validator API creates a layer of indirection when
testing the behaviour of the validator. Ideally, the validator could
be given any API to promote understandability and readability.

This change allows for validators to be dynamically created with any API
the developer desires.

An example showing the change in readability and flexibility:

``` ruby
class PolicyValidator
  attr_reader :desired_policy, :allowed_policies
  def initialize(desired_policy:, allowed_policies:)
    @desired_policy = desired_policy
    @allowed_policies = allowed_policies
  end

  def call(errors)
    unless allowed_policies.include?(desired_policy)
      errors.add(:desired_policy, "the desired policy is not allowed")
    end
  end
end

class AdminUserRequest
  include Valcro

  attr_accessor :desired_policy

  validates_with do
    PolicyValidator.new(desired_policy: desired_policy,
                        allowed_policies: ["admin"])
  end
end

class StandardUserRequest
  include Valcro

  attr_accessor :desired_policy, :policy_service

  validates_with do
    PolicyValidator.new(desired_policy: desired_policy,
                        allowed_policies: policy_service.call)
  end
end
```